Add files via upload
This is effectively (R_BY-R_RM) estimated from top 20% and bottom 20% of growth rate for each SNP in each environment. If we take abs() of this value, we result in sH and sL. This result is slightly "rawer", which allows testing diminishing returns by constraining to the same sign or ignore the sign of effect. The QTL (annotated by the SNP index) effect could be extracted from this table.